Updated all `astype` calls to use np types

+++ stumpy/aampi.py
@@ -97,9 +97,9 @@ class aampi:
         self._egress = egress
 
         mp = aamp(self._T, self._m)
-        self._P = mp[:, 0]
-        self._I = mp[:, 1]
-        self._left_I = mp[:, 2]
+        self._P = mp[:, 0].astype(np.float64)
+        self._I = mp[:, 1].astype(np.int64)
+        self._left_I = mp[:, 2].astype(np.int64)
         self._left_P = np.empty(self._P.shape)
         self._left_P[:] = np.inf
